refactor: address PR review - decouple service from FastAPI
- Move llm_completion_error_to_http_exception to routers/web_utils.py
- Remove FastAPI dependencies from services/llm_service.py
- Simplify error handling by catching httpx exceptions directly
- Remove redundant exception re-raise blocks
- Document explicit exception handling approach in chat.py

Addresses PR #70 review comments:
- Service layer no longer coupled to FastAPI (HTTPException moved to web tier)
- Error handling simplified (httpx exceptions bubble up naturally)
- Redundant re-throw blocks removed
- Global exception handler deferred to Step 6 per incremental approach
